Refactored deck building logic into cardService
