--- src/main/java/com/blackjack/models/CardGraphics.java
@@ -1,8 +1,0 @@
-package com.blackjack.models;
-
-import com.blackjack.enums.Rank;
-import com.blackjack.enums.Suit;
-
-public class CardGraphics {
-	
-}
